Remove wget in favour of cURL
I'm proposing removing the use of wget for two reasons:

 - Using wget was set to `--quiet` which hides an error message around
using the `--no-check-certificate` but users can't see that. cURL works
and doesn't need to have any extra flags passed to the command.

 - wget is not installed as standard on macs where cURL is, alternatively
I'm happy for this to be declined and add an error on it being missing as
a pre-install check? This is possibly causing issues for others installing
on the latest versions of macOS too.